requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
async-generator==1.10
attrs==21.4.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.1.0
cryptography==37.0.2
glob2==0.7
h11==0.13.0
idna==3.3
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
outcome==1.1.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
pluggy==1.0.0
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-bdd==6.0.1
python-dotenv==0.20.0
requests==2.28.1
selenium==4.2.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
